move other comp to it own class except for custodian and author
